requirements.txt
numpy==1.17.2
tqdm==4.36.1
torchvision==0.6.0
torch==1.5.0
Pillow==7.1.2
sklearn